Added initial font_renderer_default.cpp
