+++ engine/render/src/test/wscript
@@ -1,7 +1,7 @@
 #! /usr/bin/env python
 
 def build(bld):
-    libs = 'TESTMAIN SOCKET APP GRAPHICS_NULL DDF SCRIPT EXTENSION RESOURCE DLIB PROFILE_NULL PLATFORM_NULL LUA render'
+    libs = 'TESTMAIN SOCKET APP GRAPHICS_NULL DDF SCRIPT EXTENSION RESOURCE DLIB PROFILE_NULL PLATFORM_NULL LUA render render_font_default'
     exported_symbols = ['GraphicsAdapterNull']
 
     if 'web' in bld.env['PLATFORM']:
